requirements.txt
# TODO: This do not contain everything yet
# I need to investigate on what to add
alembic==1.10.3
aniso8601==9.0.1
bcrypt==4.0.1
black==23.3.0
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
Flask-Bcrypt==1.0.1
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
psycopg2-binary==2.9.6
PyJWT==2.6.0
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2023.3
requests==2.28.2
sendgrid==6.10.0
six==1.16.0
SQLAlchemy==2.0.9
starkbank-ecdsa==2.2.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
